feat: decimal consumable quantities (#98 discussion)
Half a can of spray is 0.5 (Discussion #98): stock, consumption links,
restock deltas and reorder thresholds accept decimals end to end, so the
CAN stays the unit (price 4.99/can) instead of a millilitre workaround
that broke on price precision.

Unit-dimension sweep (the #58/#59 audit checklist — every consumer):
- helpers/parts: round_qty() = canonical 2-decimal rounding, whole numbers
  collapse to int; _clean_stock/floats; sanitize_consumes_parts keeps 0.5,
  zero/invalid still falls back to 1; part_is_low compares floats
- storage: set/get_part_stock float, rounded at the single write choke point
- parts_runtime: float deltas/absolutes/quantities in the consume + restock
  + buy-task math
- WS schemas: part stock/restock_quantity/delta and the complete-dialog
  quantity-bought accept numbers (Coerce(float) with ranges)
- io import: float stock survives the JSON round-trip
- frontend: parseFloat + step=0.01 on restock input, part form (Number()
  already float-capable), complete-dialog quantity fields, task-dialog
  consumes-quantity input

Tests: #98 scenario pinned (0.5 consumption round-trip, decimal stock,
fractional low-threshold comparison, rounding collapse); full backend
2,775 @ gate; frontend suite green; ruff + mypy clean.

+++ custom_components/maintenance_supporter/parts_runtime.py
@@ -55,7 +55,7 @@ def _fire_transition(
     hass: HomeAssistant,
     entry: ConfigEntry,
     part: dict[str, Any],
-    stock: int | None,
+    stock: float | None,
     transition: str | None,
 ) -> None:
     event = _TRANSITION_EVENTS.get(transition or "")
@@ -87,9 +87,9 @@ async def async_change_part_stock(
     entry: ConfigEntry,
     part_id: str,
     *,
-    delta: int | None = None,
+    delta: float | None = None,
     absolute: int | None = None,
-) -> int | None:
+) -> float | None:
     """Change one part's stock (clamped at 0), fire the edge event, save.
 
     Returns the new stock, or None when the part is unknown / untracked with a
@@ -102,9 +102,9 @@ async def async_change_part_stock(
         return None
     old = store.get_part_stock(part_id)
     if absolute is not None:
-        new = max(0, int(absolute))
+        new = max(0.0, float(absolute))
     else:
-        new = max(0, (old or 0) + int(delta or 0))
+        new = max(0.0, (old or 0) + float(delta or 0))
     store.set_part_stock(part_id, new)
     # Immediate save: part CRUD and the buy-task reconcile may reload the entry
     # right after, which re-reads the store from disk — a debounced save would
@@ -121,7 +121,7 @@ async def async_handle_completion_parts(
     entry: ConfigEntry,
     task_data: dict[str, Any],
     *,
-    restock_quantity: int | None = None,
+    restock_quantity: float | None = None,
 ) -> None:
     """Completion-side part effects: consume linked parts / restock a buy task.
 
@@ -140,7 +140,7 @@ async def async_handle_completion_parts(
     ref = task_data.get(PART_REF_FIELD)
     if isinstance(ref, dict) and ref.get("part_id") in parts:
         part = parts[ref["part_id"]]
-        qty = restock_quantity if restock_quantity and restock_quantity > 0 else int(part.get("restock_quantity") or 1)
+        qty = restock_quantity if restock_quantity and restock_quantity > 0 else float(part.get("restock_quantity") or 1)
         old = store.get_part_stock(part["id"])
         new = max(0, (old or 0) + qty)
         store.set_part_stock(part["id"], new)
@@ -156,7 +156,7 @@ async def async_handle_completion_parts(
         old = store.get_part_stock(part["id"])
         if old is None:
             continue  # catalog-only part — nothing to decrement
-        qty = int(link.get("quantity", 1) or 1)
+        qty = float(link.get("quantity", 1) or 1)
         new = max(0, old - qty)
         store.set_part_stock(part["id"], new)
         _fire_transition(hass, entry, part, new, stock_transition(part, old, new))
